utils reorganization
